appendix graphs and a few formulation changes

+++ Code/Simple/UnalignedNaivePrecomputed/Node.h
@@ -22,7 +22,7 @@
 #include <assert.h>
 
 typedef std::vector<bool> bitmap_t;
-typedef std::vector<ushort> blockRanksVector;
+typedef std::vector<uint> blockRanksVector;
 
 using namespace std;
